Add the -Z html-output option and ui header command
Our goal is to generate HTML pages representing the error emitted by the
compiler.

This commit adds a -Z html-output flag to rustc. When passed, the
compiler will emit its errors in HTML format in stderr. This flag
conflicts with any value of the error-format argument.

As a bonus, the compiletest crate has been updated to handle
the // html-output compiler directive. Adding this flag will invoke the
compiler with the -Z html-output flag and compare its output with a
corresponding .html file, similarly as .stderr files.